apps/prompt_lab/requirements.txt
flask>=3.1.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
httpx>=0.27.0
python-frontmatter>=1.1.0
